Fix a typo
git-svn-id: https://local.ugene.unipro.ru/svn/ugene/trunk@3584 7ca3a678-163c-e011-b84f-001e8c0267ff

+++ src/plugins/CoreTests/src/DocumentModelTests.cpp
@@ -228,7 +228,9 @@ Task::ReportResult GTest_LoadBrokenDocument::report() {
 }
 
 void GTest_LoadBrokenDocument::cleanup() {
-    QFile::remove(url);
+    if (tempFile) {
+        QFile::remove(url);
+    }
 }
 
 /*******************************
